document_extraction_system (copy)/requirements.txt
# requirements.txt
mistralai>=0.1.0
langchain>=0.1.0
langchain-mistralai>=0.0.1
langchain-core>=0.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
pathlib
typing-extensions>=4.0.0
pip install -U langchain-community